testing-v2/scenarios/gaming-leaderboard/iterations/iteration-003-python/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
azure-cosmos>=4.7.0
aiohttp>=3.13.3
pydantic>=2.9.0
